move: regression models into src/prediction/regression/**
