+++ src/prediction/cnn/data_loader.py
@@ -1,0 +1,4 @@
+# Load data for CNN model
+# Encode Grid coord into image N x N
+class DataLoader:
+    pass
