+++ src/prediction/regression/trainers/ada_model_trainer.py
@@ -4,7 +4,7 @@ ModelTrainer class which is responsible for training a AdaBoostRegressor (ADA) m
 
 from sklearn.ensemble import AdaBoostRegressor
 
-from src.prediction.model_trainer import ModelTrainer
+from src.prediction.regression.regression_model_trainer import ModelTrainer
 
 # pylint: disable=invalid-name
