+++ src/prediction/regression/trainers/rf_model_trainer.py
@@ -4,7 +4,8 @@ ModelTrainer class which is responsible for training a random Forest Regressor (
 
 from sklearn.ensemble import RandomForestRegressor
 
-from src.prediction.model_trainer import ModelTrainer
+from src.prediction.regression.regression_model_trainer import ModelTrainer
+
 
 # pylint: disable=invalid-name
